BUG NoDependencyBenchmark: Removed thread approach from benchmark
Reason: The threads had been executed sequentially because Thread.run() instead of Thread.start() had been called. With the correct method call, the thread approach was significantly faster than the other two approaches for certain cases and we could not find a satisfying explanation for this. Since we want to benchmark the performance of Hippodamus primarily against CompletableFutures and this benchmark was the only one considering the thread approach, we decided to remove the thread approach entirely.